requirements.txt
pytest
pytest-cov
pyyaml
logging_tree
flake8
twine
